robustify code: - remove some superfluous checks - prevent overflow where range of p sets is calculated - use 64bit lattice size threshold in GPU core (maybe not currently needed, but it doesn't hurt)
git-svn-id: https://svn.code.sf.net/p/msieve/code/trunk@481 8c02911e-0617-4013-80d0-c3bcbcae04d5

+++ gnfs/poly/stage1/stage1_core_gpu/stage1_core.h
@@ -44,7 +44,7 @@ typedef struct {
 
 typedef struct {
 	uint32 p[P_SOA_BATCH_SIZE];
-	uint32 lattice_size[P_SOA_BATCH_SIZE];
+	uint64 lattice_size[P_SOA_BATCH_SIZE];
 	uint64 roots[POLY_BATCH_SIZE][P_SOA_BATCH_SIZE];
 } p_soa_t;
